pyproject.toml
[tool.poetry]
name = "diploma"
version = "0.1.0"
description = ""
authors = ["verghil <verghil@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "~4.2.3"
requests = "~2.31.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
